Move more items to xelnaga.
