+++ Pylon/Services/UrlConverter.cs
@@ -1,6 +1,6 @@
-﻿using Aiursoft.Pylon.Interfaces;
-using Aiursoft.Pylon.Models;
-using Aiursoft.Pylon.Models.API.OAuthAddressModels;
+﻿using Aiursoft.Pylon.Models.API.OAuthAddressModels;
+using Aiursoft.XelNaga.Interfaces;
+using Aiursoft.XelNaga.Models;
 
 namespace Aiursoft.Pylon.Services
 {
