+++ Pylon/Services/WebSocketExtends.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Net.WebSockets;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Aiursoft.Pylon.Services
+{
+    public static class WebSocketExtends
+    {
+        public static async Task SendMessage(this WebSocket ws, string message)
+        {
+            var buffer = new ArraySegment<byte>(Encoding.UTF8.GetBytes(message));
+            await ws.SendAsync(buffer, WebSocketMessageType.Text, true, CancellationToken.None);
+        }
+    }
+}
