+++ XelNaga/Models/AiurKeyPair.cs
@@ -1,5 +1,5 @@
 ﻿using Aiursoft.XelNaga.Interfaces;
-using Aiursoft.XelNaga.Services;
+using Aiursoft.XelNaga.Tools;
 using Microsoft.Extensions.Configuration;
 using System.Security.Cryptography;
