+++ XelNaga/Models/AiurProtocal.cs
@@ -1,5 +1,5 @@
 ﻿using Aiursoft.XelNaga.Interfaces;
-using Aiursoft.XelNaga.Services;
+using Aiursoft.XelNaga.Tools;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
