+++ XelNaga/Models/AiurUrl.cs
@@ -1,4 +1,4 @@
-﻿using Aiursoft.XelNaga.Services;
+﻿using Aiursoft.XelNaga.Tools;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
