+++ XelNaga/Services/RSAService.cs
@@ -1,5 +1,6 @@
 ﻿using Aiursoft.XelNaga.Interfaces;
 using Aiursoft.XelNaga.Models;
+using Aiursoft.XelNaga.Tools;
 using System.Security.Cryptography;
 
 namespace Aiursoft.XelNaga.Services
